Add partcipants to rooms

+++ dreams/views.py
@@ -1,17 +1,23 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
-from .models import Room, Topic
+from django.contrib.auth.forms import UserCreationForm
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 # Create your views here.
 
 def login_page(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('dreams:home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -27,12 +33,28 @@ def login_page(request):
         else:
             messages.error(request, 'Username or password does not exist')
 
-    context = {}
+    context = {'page': page}
     return render(request, 'dreams/login_register.html', context)
 
 def logout_user(request):
     logout(request)
     return redirect('dreams:home')
+
+def register_page(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit = False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('dreams:home')
+        else:
+            messages.error(request, 'An error ocurred during registration')
+
+    return render(request, 'dreams/login_register.html', {'form': form})
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -54,7 +76,21 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id = pk)
-    context = {'room': room}
+    room_messages = room.message_set.all().order_by('created')
+    participants = room.participants.all()
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        
+        room.participants.add(request.user)
+
+        return redirect('dreams:room', pk = room.id)
+
+    context = {'room': room, 'room_messages': room_messages, 'participants': participants}
     
     return render(request, 'dreams/room.html', context)
 
@@ -75,6 +111,9 @@ def create_room(request):
 def update_room(request, pk):
     room = Room.objects.get(id = pk)
     form = RoomForm(instance = room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
@@ -90,10 +129,24 @@ def update_room(request, pk):
 def delete_room(request, pk):
     room = Room.objects.get(id = pk)
     
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('dreams:home')
 
     return render(request, 'dreams/delete.html', {'obj': room})
 
+@login_required(login_url='dreams:login')
+def delete_message(request, pk):
+    message = Message.objects.get(id = pk)
+    
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here')
+    
+    if request.method == 'POST':
+        message.delete()
+        return redirect('dreams:home')
 
+    return render(request, 'dreams/delete.html', {'obj': message})
